+++ src/main/java/au/vic/delwp/IDnText.java
@@ -2,6 +2,6 @@ package au.gov.vic.delwp;
 
 public class IDnText{
 
-	public short ID;
+	public int ID;
 	public String Text;
 	}
